Авторизация в разделе Роли

+++ app/Http/Controllers/RoleController.php
@@ -3,16 +3,23 @@
 namespace Chunker\Base\Http\Controllers;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Request;
 use Chunker\Base\Http\Requests\RoleRequest;
 use Chunker\Base\Models\Role;
+use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
 	/*
 	 * Страница роли
 	 */
-	public function form(Role $role = NULL) {
+	public function form(Request $request, Role $role = NULL) {
+		$this->authorize('roles.view');
+
+		// Если пользователь не может редактировать, то и создавать не может
+		if ($request->user()->cannot('roles.edit') && !$role->exists) {
+			return redirect()->route('admin.roles', Role::orderBy('name')->first());
+		}
+
 		// Роли
 		$_roles = Role
 			::orderBy('name')
@@ -35,6 +42,8 @@ class RoleController extends Controller
 	 * Добавление роли
 	 */
 	public function store(RoleRequest $request) {
+		$this->authorize('roles.edit');
+
 		$role = Role::create($request->all());
 		$this->syncAbilities($request, $role);
 		flash()->success('Роль <b>' . $role->name . '</b> добавлена');
@@ -47,6 +56,8 @@ class RoleController extends Controller
 	 * Обновление роли
 	 */
 	public function update(RoleRequest $request, Role $role) {
+		$this->authorize('roles.edit');
+
 		$role->update($request->all());
 		$this->syncAbilities($request, $role);
 		flash()->success('Данные роли <b>' . $role->name . '</b> сохранены');
@@ -59,6 +70,8 @@ class RoleController extends Controller
 	 * Удаление роли
 	 */
 	public function destroy(Role $role) {
+		$this->authorize('roles.edit');
+
 		$role->delete();
 		flash()->warning('Роль <b>' . $role->name . '</b> удалена');
 
@@ -75,15 +88,15 @@ class RoleController extends Controller
 
 			foreach ($request->get('abilities') as $namespace => $ability) {
 				// Отсев невозможного
-				if ($ability) {
+				if ($ability && $request->user()->can($namespace . '.edit')) {
 					$abilities[] = $ability;
 				}
-
-				// Синхронизация
-				$role
-					->abilities()
-					->sync($abilities);
 			}
+
+			// Синхронизация
+			$role
+				->abilities()
+				->sync($abilities);
 		}
 	}
 }
